rl_code/requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
black==22.8.0
catboost==1.0.6
chex==0.1.4
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.7
etils==0.7.1
executing==1.0.0
flax==0.6.0
fonttools==4.37.1
gym==0.25.2
gym-notices==0.0.8
hydra-core==1.2.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipdb==0.13.9
ipython==8.4.0
# jax==0.3.16
# jaxlib==0.3.15
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
msgpack==1.0.4
numpy==1.23.2
omegaconf==2.2.3
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
rich==11.2.0
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
stack-data==0.5.0
tinygp==0.2.2
toml==0.10.2
toolz==0.12.0
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
wcwidth==0.2.5
zipp==3.8.1
hydra-joblib-launcher==1.2.0
